ansi_code_processor: split out qt-unrelated part
